+++ app/components/about/about.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
 import myGlobals = require('../../globals'); 
+import { Title }     from '@angular/platform-browser';
+
 
 @Component({
   templateUrl: 'themes/' + myGlobals.themeName + '/about.tpl.html',
@@ -7,4 +9,12 @@ import myGlobals = require('../../globals');
 
 export class AboutComponent {
 	public pageHeading = "About";
+	public constructor(private titleService: Title ) { 
+		this.titleService.setTitle( this.pageHeading );
+	}
 }
+
+
+
+
+