+++ app/components/home/home.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { Title }     from '@angular/platform-browser';
 import myGlobals = require('../../globals'); 
 @Component({
   templateUrl: 'themes/' + myGlobals.themeName + '/home.tpl.html'
@@ -6,4 +7,9 @@ import myGlobals = require('../../globals');
 
 export class HomeComponent {
 	public pageHeading = "Welcome";
+	public constructor(private titleService: Title ) { 
+		this.titleService.setTitle( this.pageHeading );
+	}
 }
+
+