+++ app/components/profile/profile.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { Title }     from '@angular/platform-browser';
 import myGlobals = require('../../globals'); 
 @Component({
   templateUrl: 'themes/' + myGlobals.themeName + '/profile.tpl.html'
@@ -6,4 +7,7 @@ import myGlobals = require('../../globals');
 
 export class ProfileComponent {
 	public pageHeading = "Profile";
+	public constructor(private titleService: Title ) { 
+		this.titleService.setTitle( this.pageHeading );
+	}
 }